move news components to feature folder
