+++ src/components/feature-news-form/index.ts
@@ -1,0 +1,3 @@
+import FormNews from "./FormNews.vue";
+
+export { FormNews };
