+++ src/pages/NewsCreate.page.vue
@@ -21,7 +21,7 @@
 
 <script>
 import DefaultWireframe from "@/components/templates/DefaultWireframe.vue";
-import FormNews from "@/components/organisms/FormNews";
+import { FormNews } from "@feature-news-form";
 import { newsModule, notifierModule } from "@/store";
 import { buildPageTitle } from "@/utils/pageTitle";
