+++ tsconfig.json
@@ -42,6 +42,7 @@
 			"@feature-board": ["src/components/feature-board"],
 			"@feature-editor": ["src/components/feature-editor"],
 			"@feature-render-html": ["src/components/feature-render-html"],
+			"@feature-news-form": ["src/components/feature-news-form"],
 			"@ui-alert": ["src/components/ui-alert"],
 			"@ui-board": ["src/components/ui-board"],
 			"@ui-color-overlay": ["src/components/ui-color-overlay"],
